cleanup update ipc code
related to #114418
